refactor: move season match insights slice into core application
