+++ backend/src/api/interface/controller/v1/season_competition_controller.py
@@ -39,10 +39,12 @@ from api.interface.controller.v1.season_competition_presenters import (
 )
 from api.middleware.exception_mapper import map_exceptions
 from auth.dependencies import authorize_pena_access, require_admin
+from core.application.use_cases.get_season_match_insights_usecase import (
+    GetSeasonMatchInsightsUseCase,
+)
 from fastapi import APIRouter, Depends, Query, status
 from persistence.application.update_policies import FieldUpdate, StandingsUpdatePolicy
 from persistence.application.use_cases import (
-    GetSeasonMatchInsightsUseCase,
     InvalidSeasonMatchDataError,
     InvalidSeasonPlayerUpdateDataError,
     ManageSeasonMatchesUseCase,
