+++ backend/src/core/application/models/__init__.py
@@ -1,0 +1,13 @@
+from .season_match_insights_models import (
+    MatchDetail,
+    MatchInsightRow,
+    MatchPlayerStats,
+    MatchTeam,
+)
+
+__all__ = [
+    "MatchDetail",
+    "MatchInsightRow",
+    "MatchPlayerStats",
+    "MatchTeam",
+]
