+++ backend/src/core/application/ports/__init__.py
@@ -1,0 +1,3 @@
+from .season_match_insights_port import SeasonMatchInsightsPort
+
+__all__ = ["SeasonMatchInsightsPort"]
