+++ backend/src/core/application/ports/season_match_insights_port.py
@@ -1,0 +1,12 @@
+from typing import Protocol
+
+from core.application.models import MatchInsightRow
+
+
+class SeasonMatchInsightsPort(Protocol):
+    def list_closed_match_insight_rows(
+        self,
+        *,
+        pena_guid: str,
+        season_guids: list[str],
+    ) -> list[MatchInsightRow]: ...
