+++ backend/src/core/application/services/__init__.py
@@ -1,0 +1,3 @@
+from .match_insights_report_builder import MatchInsightsAccumulator, MatchInsightsReportBuilder
+
+__all__ = ["MatchInsightsAccumulator", "MatchInsightsReportBuilder"]
