+++ backend/src/core/application/use_cases/season_match_insights_errors.py
@@ -1,0 +1,10 @@
+class InvalidSeasonInsightsDataError(Exception):
+    pass
+
+
+class PenaSeasonPenaNotFoundError(Exception):
+    pass
+
+
+class PenaSeasonNotFoundError(Exception):
+    pass
