+++ backend/src/persistence/application/services/__init__.py
@@ -1,3 +1,3 @@
-from .match_insights_report_builder import MatchInsightsReportBuilder
+from core.application.services.match_insights_report_builder import MatchInsightsReportBuilder
 
 __all__ = ["MatchInsightsReportBuilder"]
